+++ src/core/client/ui/components/RelativeTime/index.ts
@@ -1,0 +1,2 @@
+export * from "./RelativeTime";
+export { default } from "./RelativeTime";
