--- src/core/client/ui/components/Timestamp/index.ts
@@ -1,2 +1,0 @@
-export * from "./Timestamp";
-export { default } from "./Timestamp";
