+++ src/core/client/ui/components/index.ts
@@ -1,7 +1,7 @@
 export { default as BaseButton } from "./BaseButton";
 export { default as Button } from "./Button";
 export { default as Typography } from "./Typography";
-export { default as Timestamp } from "./Timestamp";
+export { default as RelativeTime } from "./RelativeTime";
 export { default as UIContext } from "./UIContext";
 export { default as Flex } from "./Flex";
 export { default as MatchMedia } from "./MatchMedia";
